+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Requires PHP 5.6 and WordPress 4.7 or above
 * Enhance: adjust styling for setup instructions (#215, props timse201)
 * Enhance: update hooks for Multisite initialization in WordPress 5.1 and above (#246, props ouun)
 * Enhance: rework flush hooks and add some third-party triggers for Autoptimize and WooCommerce (#225, props timse201)
-
+* Fix: correctly add user-agent to robots.txt (#282) (#283)
 
 ## 2.3.2
 * Fix: enforce WordPress environment for caching modules (#221, props timse201)